Order Management System (OMS)/requirements.txt
sqlalchemy
python-dotenv
customtkinter
pillow
matplotlib
bcrypt
